Enable SCMG message definitions
Even if the SCCP_Emulation internally doesn't itself support SCMG, there
is a lot of use in having the SCMG message/type definitions enabled
(they are currently commented out).

SCMG operates as a SCCP user on SSN=1, so having those definitions
available allows a TTCN3 test case to use the message definitions to
implement SSN=1 on top of SCCP_Emulation.

Signed-off-by: Harald Welte <laforge@gnumonks.org>
Change-Id: I176b63542cbf844a2b86e01ec2534b76cf3c2f6b